Rename the `anoma node` command to `anoma start`
