+++ app/Commands/Dev/Anoma.hs
@@ -5,9 +5,9 @@ module Commands.Dev.Anoma
 where
 
 import Commands.Base
-import Commands.Dev.Anoma.Node qualified as Node
 import Commands.Dev.Anoma.Options
+import Commands.Dev.Anoma.Start qualified as Start
 
 runCommand :: (Members AppEffects r) => AnomaCommand -> Sem r ()
 runCommand = \case
-  AnomaCommandNode opts -> Node.runCommand opts
+  AnomaCommandStart opts -> Start.runCommand opts
